Risk management module

+++ src/MarginTrading.AzureRepositories/MarginTradingPositionEntity.cs
@@ -3,32 +3,32 @@ using Microsoft.WindowsAzure.Storage.Table;
 
 namespace MarginTrading.AzureRepositories
 {
-    public class MarginTradingPositionEntity : TableEntity, IPosition
-    {
-        public string ClientId { get; set; }
-        public string Asset { get; set; }
-        public decimal Volume { get; set; }
+	public class MarginTradingPositionEntity : TableEntity, IPosition
+	{
+		public string ClientId { get; set; }
+		public string Asset { get; set; }
+		public double Volume { get; set; }
 
-        public static string GeneratePartitionKey(string asset)
-        {
-            return asset;
-        }
+		public static string GeneratePartitionKey(string asset)
+		{
+			return asset;
+		}
 
-        public static string GenerateRowKey(string clientId, string asset)
-        {
-            return $"{clientId}_{asset}";
-        }
+		public static string GenerateRowKey(string clientId, string asset)
+		{
+			return $"{clientId}_{asset}";
+		}
 
-        public static MarginTradingPositionEntity Create(IPosition position)
-        {
-            return new MarginTradingPositionEntity
-            {
-                PartitionKey = GeneratePartitionKey(position.Asset),
-                RowKey = GenerateRowKey(position.ClientId, position.Asset),
-                Asset = position.Asset,
-                ClientId = position.ClientId,
-                Volume = position.Volume
-            };
-        }
-    }
+		public static MarginTradingPositionEntity Create(IPosition position)
+		{
+			return new MarginTradingPositionEntity
+			{
+				PartitionKey = GeneratePartitionKey(position.Asset),
+				RowKey = GenerateRowKey(position.ClientId, position.Asset),
+				Asset = position.Asset,
+				ClientId = position.ClientId,
+				Volume = position.Volume
+			};
+		}
+	}
 }
