+++ src/MarginTrading.Core/IPosition.cs
@@ -1,9 +1,9 @@
 ﻿namespace MarginTrading.Core
 {
-    public interface IPosition
-    {
-        string ClientId { get; set; }
-        string Asset { get; set; }
-        decimal Volume { get; set; }
-    }
+	public interface IPosition
+	{
+		string ClientId { get; set; }
+		string Asset { get; set; }
+		double Volume { get; set; }
+	}
 }
