+++ src/MarginTrading.Core/Position.cs
@@ -1,9 +1,9 @@
 ﻿namespace MarginTrading.Core
 {
-    public class Position : IPosition
-    {
-        public string ClientId { get; set; }
-        public string Asset { get; set; }
-        public decimal Volume { get; set; }
-    }
+	public class Position : IPosition
+	{
+		public string ClientId { get; set; }
+		public string Asset { get; set; }
+		public double Volume { get; set; }
+	}
 }
